refactor: rename page file and dedicated test
